% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_fill_zeros.R
\name{acti_fill_zeros}
\alias{acti_fill_zeros}
\alias{acti_fill_zeroes}
\alias{acti_remove_leading_zeros}
\alias{acti_remove_leading_zeroes}
\title{Fill in Zeros}
\usage{
acti_fill_zeros(data)

acti_fill_zeroes(data)

acti_remove_leading_zeros(data)

acti_remove_leading_zeroes(data)
}
\arguments{
\item{data}{a data frame containing columns time, X, Y, Z}
}
\value{
the modified data frame with zeros replaced by NA
}
\description{
Fill in Zeros
}
\examples{
acti_fill_zeros(acti_raw_data)
acti_fill_zeroes(acti_raw_data)
}
