% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_separate_times.R
\name{acti_separate_time}
\alias{acti_separate_time}
\alias{acti_separate_times}
\alias{acti_create_date}
\alias{acti_create_hour}
\alias{acti_create_minute}
\title{Separate Times into Date, Hour, and Minute}
\usage{
acti_separate_time(data)

acti_separate_times(data)

acti_create_date(data)

acti_create_hour(data)

acti_create_minute(data)
}
\arguments{
\item{data}{a \code{data.frame} with a \code{time} column}
}
\value{
A \code{data.frame} with date, hour, minute, and day columns
}
\description{
Separate Times into Date, Hour, and Minute
}
\examples{
library(actibase)
acti_separate_time(acti_raw_data)
acti_create_date(acti_raw_data)
acti_create_hour(acti_raw_data)
acti_create_minute(acti_raw_data)
}
