% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_axes.R
\name{acti_standardize_data}
\alias{acti_standardize_data}
\alias{acti_standardise_data}
\title{Standardize the Accelerometry Data}
\usage{
acti_standardize_data(
  data,
  subset_xyz = TRUE,
  colname_time = "time",
  lower_case = FALSE,
  check_xyz = TRUE
)

acti_standardise_data(
  data,
  subset_xyz = TRUE,
  colname_time = "time",
  lower_case = FALSE,
  check_xyz = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} with a column for time in \code{POSIXct} (usually
\code{time}), and \code{X}, \code{Y}, \code{Z}}

\item{subset_xyz}{should only the \code{time} (if available)
and \code{XYZ} be subset?}

\item{colname_time}{column name of header for time}

\item{lower_case}{flag to make sure all columns are lower case.}

\item{check_xyz}{Check if X/Y/Z is in the data}
}
\value{
A \code{data.frame} with \code{X/Y/Z} and a time in
\code{time} (if available).
}
\description{
Standardize the Accelerometry Data
}
\examples{

acti_standardize_data(acti_raw_data)
acti_standardise_data(acti_raw_data)

}
