% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_axes.R
\name{acti_tidy_axes}
\alias{acti_tidy_axes}
\title{Tidy axes to a long format}
\usage{
acti_tidy_axes(data, colname_time = "time")
}
\arguments{
\item{data}{An object with columns a time column \code{X}, \code{Y}, and \code{Z} or an
object of class \code{AccData}}

\item{colname_time}{column name of header for time}
}
\value{
A long data set with \code{time}, \code{axis}, and \code{value}
}
\description{
Tidy axes to a long format
}
\examples{
long = acti_tidy_axes(acti_raw_data)
}
