% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_day_inclusion.R
\name{create_day_inclusion}
\alias{create_day_inclusion}
\alias{add_day_inclusion}
\title{Create Day-Level Inclusion information}
\usage{
create_day_inclusion(data, min_required = 1368L)

add_day_inclusion(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} with the columns \code{time}}

\item{min_required}{Number of minutes required in a day to be called \code{included}}

\item{...}{arguments to pass to \link{create_day_inclusion} when using
\link{add_day_inclusion}}
}
\value{
A \code{data.frame} for each day with information of number of minutes observed
and included
}
\description{
Create Day-Level Inclusion information
}
\examples{
data = acti_raw_data |>
  dplyr::mutate(r = sqrt(X^2 + Y^2 + Z^2),
                time = lubridate::floor_date(time, "1 minute")) |>
  dplyr::group_by(time) |>
  dplyr::summarise(r = sum(r), .groups = "drop") |>
  dplyr::mutate(wear = r > 4000)

res = create_day_inclusion(data)
}
