% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_minute.R
\name{hms_to_minute}
\alias{hms_to_minute}
\title{Convert an `hms` Time to a Minute of the Day}
\usage{
hms_to_minute(x, start = 1L)
}
\arguments{
\item{x}{A vector of class `hms`.}

\item{start}{The index assigned to midnight; either `0` or `1`.}
}
\value{
An integer vector with values from `start` through `start + 1439`.
}
\description{
Converts a time of day to its minute position within a 24-hour day. By
default, positions are one-based, so midnight is minute 1 and 23:59 is
minute 1440.
}
\examples{
hms_to_minute(hms::as_hms(c("00:00:00", "23:59:00")))
hms_to_minute(hms::as_hms("00:00:00"), start = 0)
}
