% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{strip_hour_shift}
\alias{strip_hour_shift}
\title{Strip Hour Shift from Character Time Vector}
\usage{
strip_hour_shift(x, max_index = 2L)
}
\arguments{
\item{x}{character vector with times that may include hour shifts
(e.g., "2019-01-01 12:00+03:00")}

\item{max_index}{maximum index to grab the shift from after splitting on
spaces, default is 2 (e.g., "2019-01-01 12:00")}
}
\value{
A character vector with the \code{+}/\code{-} hour shift removed
}
\description{
Strip Hour Shift from Character Time Vector
}
\examples{
strip_hour_shift(c("2019-01-01 12:00+03:00", "2019-01-01 12:00-04:00"))
}
