% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_raw_data.R
\docType{data}
\name{acti_raw_data}
\alias{acti_raw_data}
\title{Example Actigraphy/Activity Raw Data}
\format{
A data.frame with the columns
\describe{
\item{time}{time column}
\item{X}{X column}
\item{Y}{Y column}
\item{Z}{Y column}
}
}
\usage{
acti_raw_data
}
\description{
Example Actigraphy/Activity Raw Data
}
\keyword{datasets}
