% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_resample.R
\name{acti_resample}
\alias{acti_resample}
\alias{acti_resample_to_time}
\title{Resample 3-axial input signal to a specific sampling rate}
\usage{
acti_resample(data, sample_rate, method = "linear", ...)

acti_resample_to_time(data, times, method = "linear", ...)
}
\arguments{
\item{data}{A `data.frame` with a column for time in `POSIXct` (usually
`time`), and `X`, `Y`, `Z`}

\item{sample_rate}{sampling frequency, coercible to an integer.
This is the sampling rate you're sampling the data *into*.}

\item{method}{method for interpolation. Options are
`"linear"/"constant"`, which uses `stats::approx`, or one of
`"fmm", "periodic", "natural", "monoH.FC", "hyman"`, which uses
`stats::spline`}

\item{...}{additional arguments to pass to [stats::approx()] or
[stats::spline]}

\item{times}{a vector of `POSIXct` date/time values to interpolate
the data to}
}
\value{
A `data.frame`/`tibble` of `time` and `X`, `Y`, `Z`.
}
\description{
Resample 3-axial input signal to a specific sampling rate
}
\examples{
options(digits.secs = 3)
x = acti_raw_data
res = acti_resample(data = x, sample_rate = 80)
res = acti_resample(data = x, sample_rate = 100)
res = acti_resample(data = x, sample_rate = 1)
res = acti_resample_to_time(
  data = x,
  times = lubridate::floor_date(x$time, unit = "1 sec"),
)
res_nat = acti_resample_to_time(
  data = x,
  times = lubridate::floor_date(x$time, unit = "1 sec"),
  method = "natural"
)
}
