% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_separate_times.R
\name{acti_separate_time}
\alias{acti_separate_time}
\title{Separate Times into Date, Hour, and Minute}
\usage{
acti_separate_time(data)
}
\arguments{
\item{data}{a `data.frame` with a `time` column}
}
\value{
A `data.frame` with date, hour, minute, and day columns
}
\description{
Separate Times into Date, Hour, and Minute
}
\examples{
acti_separate_time(acti_raw_data)
}
