% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_axes.R
\name{acti_standardize_data}
\alias{acti_standardize_data}
\alias{acti_standardise_data}
\title{Standardize the Accelerometry Data}
\usage{
acti_standardize_data(
  data,
  subset_xyz = TRUE,
  colname_time = "time",
  check_xyz = TRUE
)

acti_standardise_data(
  data,
  subset_xyz = TRUE,
  colname_time = "time",
  check_xyz = TRUE
)
}
\arguments{
\item{data}{A `data.frame` with a column for time in `POSIXct` (usually
`time`), and `X`, `Y`, `Z`}

\item{subset_xyz}{should only the `time` (if available)
and `XYZ` be subset?}

\item{colname_time}{column name of header for time}

\item{check_xyz}{Check if X/Y/Z is in the data}
}
\value{
A `data.frame` with `X/Y/Z` and a time in
`time` (if available).
}
\description{
Standardize the Accelerometry Data
}
\examples{

acti_standardize_data(acti_raw_data)
acti_standardise_data(acti_raw_data)

}
