% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_minute.R
\name{acti_time_to_minute}
\alias{acti_time_to_minute}
\title{Convert a Time of Day to a Minute Index}
\usage{
acti_time_to_minute(x, start = 1L, timezone = NULL)
}
\arguments{
\item{x}{An `hms`, `POSIXt`, or `HH:MM[:SS]` character vector.}

\item{start}{The index assigned to midnight, either `0` or `1`.}

\item{timezone}{Optional Olson timezone for `POSIXt` values.}
}
\value{
An integer minute-of-day index.
}
\description{
Convert a Time of Day to a Minute Index
}
\note{
This converts each minute of a day to either `0` through `1439` or
  `1` through `1440`, the two common minute-of-day conventions.
}
\examples{
acti_time_to_minute(c("00:00", "12:00", "23:59"))
acti_time_to_minute("00:00", start = 0)
}
