% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_converters.R
\name{as_convert_safe}
\alias{as_convert_safe}
\alias{as_date_safe}
\alias{as_datetime_safe}
\title{Convert vectors ensuring no new NA}
\usage{
as_convert_safe(x, ..., func = lubridate::as_datetime)

as_date_safe(x, ...)

as_datetime_safe(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{additional arguments to pass to `func`}

\item{func}{the function to use to transform the vector `x`}
}
\value{
A converted `vector` the same length as `x` or errors if there
are introduced NAs.
}
\description{
Convert vectors ensuring no new NA
}
