% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_rate.R
\name{get_sample_rate}
\alias{get_sample_rate}
\title{Get Sample Rate}
\usage{
get_sample_rate(data, sample_rate = NULL)
}
\arguments{
\item{data}{A data set of actigraphy/activity data}

\item{sample_rate}{the sample rate.  If this is not \code{NULL}, then
it will be guess from the header or the data or the data separation}
}
\value{
A length-1 numeric vector
}
\description{
Get Sample Rate
}
