% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get_transformations}
\alias{get_transformations}
\alias{prefix_transformations}
\alias{set_transformations}
\title{Get Transformations}
\usage{
get_transformations(data)

prefix_transformations(transformations, prefix = NULL)

set_transformations(data, transformations, add = TRUE, prefix = NULL)
}
\arguments{
\item{data}{data set of data, usually time and X/Y/Z.}

\item{transformations}{character string of transformations}

\item{prefix}{if not `NULL`, the prefix plus `:` would be pasted to the
transformations.}

\item{add}{Add the transformations to those already there in `data`}
}
\value{
\code{\link{set_transformations}} returns the data, with the
`transformations` attribute updated and \code{\link{set_transformations}}
returns the attribute `transformations`
}
\description{
Get Transformations
}
