% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{is.AccData}
\alias{is.AccData}
\title{Is the object of class `AccData`}
\usage{
is.AccData(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical(1)
}
\description{
Is the object of class `AccData`
}
\examples{
is.AccData(mtcars)
}
